CHANNELMGR-1118 Fix auto-scroll for all browsers
Fixed vertical auto-scrolling for all browsers. Chrome, Safari and IE11 all support the mouse-enter/mouse-leave event on an iframe, Firefox however stops firing these events when something is beging dragged so a different, more bloated solution with attaching a mouse-move event handler *inside* the iframe is used for FF.

The scrolling itself is now done using jquery-animate on the scrollTop attribute of the html,body elements.