CHANNELMGR-1118 Cancel drag when destroying dragula
When dragging a component out of the iframe, then releasing the mouse button and clicking the 'discard changes', the page is reloaded and dragula will be destroyed. In this case we need to explicitly cancel the drag so that dragula can clean up. Failing to do so leads to different errors in different browsers.