CHANNELMGR-1149 Only use jQuery in replaceDOM function
The mix of jQuery and plain DOM manipulation led to strange
results when replaceDOM was called twice quickly. The first
manipulation was probably cached by jQuery and not yet attached
to the DOM, which made the DOM manipulation in the second call
fail on parentNode.removeChild() because parentNode was null.

Also added a unit test to cover this scenario.