CHANNELMGR-1163 Fix broken 'edit-content' button
The behavior for clicking 'edit-content' buttons should be delegated to the channel.controller in the same way as is done for the 'edit-menu' buttons.

The channel.controller.editContent function is bound to the onEditContent function of the hippo-iframe which in turns registers a callback handler with the overlay service to trigger button clicks.

I also initialized both the menu&content callback handlers in the overlay-service with angular.noop (plus tests) and removed the openContent+openMenuEditor functions from hippoIframe.controller as they are no longer used.