CHANNELMGR-1204 Avoid parallel MutationObservers in unit test
The OverlayService.init() call registers the onload handler of
the iframe, which in turn registers a MutationObserver. The init
used to be done each time the iframe was reloaded, which caused
multiple load calls and hence multiple MutationObservers. These
started observing each other's changes, which in Chrome 57 led to
an endless stream of mutations (caused by setting the label text,
which is a mutation that always reaches the DOM).

With this fixed, only one test still failed. That's because it
spied upon the _onUnload method without calling it, which
prevented the current observer from being disconnected. That
again caused two parallel observers in Chrome 57.

(cherry picked from commit e8b0ea8256f5afa0cda2a136dea6ac186a4c3717)