CHANNELMGR-1193 Fix check for hst.nomarkup xtype in isEmptyInDom
The utility method isXTypeNoMarkup checks the hst.nomarkup xtype
case-insensitive, but the isEmptyInDom method did its own comparison.
As a result a container of type hst.nomarkup (using other casing for
the xtype that the old 'HST.NoMarkup' value) thought is was empty
while is was in fact not. Since containers are rendered on top of
components and empty containers catch pointer events, the result
was the components in such a no-markup container could no longer
be clicked.

Fixed by using the utility method. Also added unit tests for this
case.