CHANNELMGR-1043 Write rich text fields
Includes 'multiple' rich text fields.

The logic to write (compound) child nodes, including the
cardinality check and the removal of remaining nodes, has been
refactored into FieldTypeUtils#writeCompoundValues. It can be
reused by CompoundFieldType, ChoiceFieldType and (new) the
RichTextFieldType.

The HTML of rich text fields is written to the hippostd:content
property as-is. The HTML is not cleaned yet, and links and images
are not handled yet either.