CHANNELMGR-1253 Show projects only in Enterprise
It is necessary to add the path to the ChannelManagerPlugin config,
because if a user does not have read rights on this projects node or the
projects tree does not exist the projects dropdown should not be
shown.

The ChannelEditor class get its config from the RootPanel and that one
in turn from the ChannelManagerPerspective. The ExtJs Wicket bridge makes
sure the properties are set on the config object. That object is loaded
into the cms service angular object that is loaded in turn in the
configservice.