CHANNELMGR-1052 Open the link picker in Visual Editing
The new link picker manager maintains the UUID of the current rich
text field compound node, so the link picker dialog can initialize
the field-specific settings. It also passes the dialog configuration
of the current rich text field to the link picker dialog.

The ckeditor component passes the field id and dialog
configuration to the backend by calling the Wicket behavior URL of
the 'StatelessLinkPickerDialogBehavior' in the link picker manager.
Since we can only pass one map of strings, the link picker
selection properties (f_uuid, f_title and f_target) and merged
with the fieldId and dialogConfig properties. This allows to
existing LinkPickerDialogBehavior to remain unchanged (just
reading the selection properties) while the
StatelessLinkPickerDialogBehavior can use the fieldId and
dialogConfig to initialize the link picker dialog for the current
rich text field before showing it. Since the dialogConfig is
actually a JSON object (with booleans, strings and arrays of
strings) it's stringified at the client and deserialized again at
the server.

Once a link is picked, the backend calls CKEditorPanel#onLickPicked,
which forwards the selected link to the ckeditor component, which
then executes the 'insertInternalLink' command from the hippopicker
plugin.