CHANNELMGR-1052 Save pending changes to HTML nodes
The pickers will create a facet node below the field node (the
returned UUID inserted in the HTML is the UUID of this facet node).
This JCR change should be saved before the Visual Editing backend
overwrites those changes again when the document is saved, otherwise
the CMS will reference deleted nodes in its pending changes.