CHANNELMGR-1037 Add FieldPath abstraction, fix writing of choices
The new FieldPath class abstracts the processing of the field path.
It consists of a first segment and all remaining segments, which
allows field types to easily parse one segment at the time.

The new method NodeFieldType#writeFieldValue is used by Choice
fields to write values directly to the chosen child node. Depending
on the type of choice there can be subtle difference between writeTo
and writeFieldValue. For example, a rich text field's writeTo
method first looks up the child node (which is named after its ID)
and writes the value into it. But with choice fields, the node
given to writeFieldValue is already the chosen child node, and the
value can be written into it directly.