CHANNELMGR-1408 Fix unit tests in Chrome 60
Change 1: forms now need to be attached to the DOM to be submitted, but
unit tests don't add compiled elements to the DOM. I'm now forcing this
in the channel settings tests by adding the compiled DOM to the Jasmine
fixtures div, so the added DOM is cleaned up automatically after each
test.

Change 2: attr('style') for an empty 'style' attribute now return an
empty string instead of undefined. Does not change the behavior of the
collapse controller, so I've adjusted the expectation in the test (which
should now work in both Chrome 59 and 60).