CHANNELMGR-1480 Add template-query service
The new REST endpoint templatequery/{id} returns a list of
DocumentTypeInfo objects {id, displayName} for the given template query.

Returns 404 if template query is not found or not of type nt:query, 500
on a query parse error or repository errors. If the targeted prototype
is of an invalid namespace or type (hippo:, hipposys:, hipposysedit:,
reporting:, nt:unstructured, hippogallery:) it will be ignored.